==> R/00_libraries.R <==
# required libraries 
install.packages("here")
install.packages("gghighlight")
install.packages("tidysynth")
install.packages("stargazer")
install.packages("tidyverse")
install.packages("tidylog")
install.packages("foreach")
devtools::install_github("ebenmichael/augsynth")
install.packages("furrr")
